User email confirmation (#36)
* Adding confirmation field to users

* Creating Blessd.Mailer

* Confirmation context and mailer

* A better new occurrence button

* Fixing format

* Making the confirmation token work properly and with expiration

* Implementing multiple churches login

* Dashboard index

* Fixing the not logged in redirection

* Finishing confirmation stuff

* Better email confirmation workflow

+++ apps/blessd/lib/blessd/memberships.ex
@@ -12,43 +12,41 @@ defmodule Blessd.Memberships do
   Returns the list of people.
   """
   def list_people(current_user) do
-    Person
-    |> Auth.check!(current_user)
-    |> Person.order()
-    |> Repo.all()
+    with {:ok, query} <- Auth.check(Person, current_user) do
+      query
+      |> Person.order()
+      |> Repo.list()
+    end
   end
 
   @doc """
   Gets a single person.
-
-  Raises `Ecto.NoResultsError` if the Person does not exist.
   """
-  def get_person!(id, current_user) do
-    Person
-    |> Auth.check!(current_user)
-    |> Repo.get!(id)
+  def find_person(id, current_user) do
+    with {:ok, query} <- Auth.check(Person, current_user), do: Repo.find(query, id)
   end
 
   @doc """
   Creates a person.
   """
   def create_person(attrs, current_user) do
-    current_user
-    |> new_person()
-    |> Person.changeset(attrs)
-    |> Repo.insert()
+    with {:ok, person} <- new_person(current_user) do
+      person
+      |> Person.changeset(attrs)
+      |> Repo.insert()
+    end
   end
 
   @doc """
   Imports a list of people from a file.
   """
   def import_people(stream, current_user) do
-    stream
-    |> CSV.decode!(headers: true)
-    |> create_people(current_user)
-    |> case do
-      {:ok, people} -> {:ok, people}
+    with enum = CSV.decode!(stream, headers: true),
+         {:ok, people} <- create_people(enum, current_user) do
+      {:ok, people}
+    else
       {:error, index, changeset} -> {:error, index + 2, changeset}
+      {:error, reason} -> {:error, reason}
     end
   end
 
@@ -59,17 +57,18 @@ defmodule Blessd.Memberships do
     enum
     |> Stream.with_index()
     |> Enum.reduce(Multi.new(), fn {row, index}, multi ->
-      changeset =
-        current_user
-        |> new_person()
-        |> Person.changeset(row)
-
-      Multi.insert(multi, index, changeset)
+      with {:ok, person} <- new_person(current_user),
+           changeset = Person.changeset(person, row) do
+        Multi.insert(multi, index, changeset)
+      else
+        {:error, reason} -> Multi.run(multi, index, fn _, _ -> {:error, reason} end)
+      end
     end)
     |> Repo.transaction()
     |> case do
       {:ok, map} -> {:ok, Map.values(map)}
-      {:error, index, changeset, _} -> {:error, index, changeset}
+      {:error, index, %Ecto.Changeset{} = changeset, _} -> {:error, index, changeset}
+      {:error, _, reason, _} -> {:error, reason}
     end
   end
 
@@ -77,34 +76,33 @@ defmodule Blessd.Memberships do
   Updates a person.
   """
   def update_person(%Person{} = person, attrs, current_user) do
-    person
-    |> Auth.check!(current_user)
-    |> Person.changeset(attrs)
-    |> Repo.update()
+    with {:ok, person} <- Auth.check(person, current_user) do
+      person
+      |> Person.changeset(attrs)
+      |> Repo.update()
+    end
   end
 
   @doc """
   Deletes a Person.
   """
   def delete_person(%Person{} = person, current_user) do
-    person
-    |> Auth.check!(current_user)
-    |> Repo.delete()
+    with {:ok, person} <- Auth.check(person, current_user), do: Repo.delete(person)
   end
 
   @doc """
   Builds a person to insert.
   """
   def new_person(current_user) do
-    Auth.check!(%Person{church_id: current_user.church.id}, current_user)
+    Auth.check(%Person{church_id: current_user.church.id}, current_user)
   end
 
   @doc """
   Returns an `%Ecto.Changeset{}` for tracking person changes.
   """
   def change_person(%Person{} = person, current_user) do
-    person
-    |> Auth.check!(current_user)
-    |> Person.changeset(%{})
+    with {:ok, person} <- Auth.check(person, current_user) do
+      {:ok, Person.changeset(person, %{})}
+    end
   end
 end
