+++ apps/blessd/lib/blessd/observance.ex
@@ -17,48 +17,49 @@ defmodule Blessd.Observance do
   Returns the list of meetings.
   """
   def list_meetings(current_user) do
-    Meeting
-    |> Auth.check!(current_user)
-    |> Meeting.preload()
-    |> Meeting.order()
-    |> Repo.all()
+    with {:ok, query} <- Auth.check(Meeting, current_user) do
+      query
+      |> Meeting.preload()
+      |> Meeting.order()
+      |> Repo.list()
+    end
   end
 
   @doc """
   Gets a single meeting.
-
-  Raises `Ecto.NoResultsError` if the Meeting does not exist.
   """
-  def get_meeting!(id, current_user) do
-    Meeting
-    |> Auth.check!(current_user)
-    |> Repo.get!(id)
+  def find_meeting(id, current_user) do
+    with {:ok, query} <- Auth.check(Meeting, current_user), do: Repo.find(query, id)
   end
 
   @doc """
   Creates a meeting.
   """
   def create_meeting(attrs, current_user) do
-    changeset =
-      current_user
-      |> new_meeting([])
-      |> Meeting.changeset(attrs)
+    with {:ok, meeting} <- new_meeting(current_user, []),
+         changeset = Meeting.changeset(meeting, attrs) do
+      Multi.new()
+      |> Multi.insert(:meeting, changeset)
+      |> Multi.run(
+        :occurrence,
+        &insert_occurrence(&1, &2, attrs["occurrences"]["0"], current_user)
+      )
+      |> Repo.transaction()
+      |> case do
+        {:ok, %{meeting: meeting, occurrence: occurrence}} ->
+          {:ok, %{meeting | occurrences: [occurrence]}}
 
-    Multi.new()
-    |> Multi.insert(:meeting, changeset)
-    |> Multi.run(:occurrence, &insert_occurrence(&1, &2, attrs["occurrences"]["0"], current_user))
-    |> Repo.transaction()
-    |> case do
-      {:ok, %{meeting: meeting, occurrence: occurrence}} ->
-        {:ok, %{meeting | occurrences: [occurrence]}}
+        {:error, :meeting, changeset, _} ->
+          {:error, changeset}
 
-      {:error, :meeting, changeset, _} ->
-        {:error, changeset}
+        {:error, assoc, %Ecto.Changeset{} = child_changeset, _} ->
+          changeset
+          |> put_assoc(assoc, child_changeset)
+          |> apply_action(:insert)
 
-      {:error, assoc, child_changeset, _} ->
-        changeset
-        |> put_assoc(assoc, child_changeset)
-        |> apply_action(:insert)
+        {:error, _, reason, _} ->
+          {:error, reason}
+      end
     end
   end
 
@@ -66,30 +67,31 @@ defmodule Blessd.Observance do
   Updates a meeting.
   """
   def update_meeting(%Meeting{} = meeting, attrs, current_user) do
-    meeting
-    |> Auth.check!(current_user)
-    |> Meeting.changeset(attrs)
-    |> Repo.update()
+    with {:ok, meeting} <- Auth.check(meeting, current_user) do
+      meeting
+      |> Meeting.changeset(attrs)
+      |> Repo.update()
+    end
   end
 
   @doc """
   Deletes a Meeting.
   """
   def delete_meeting(%Meeting{} = meeting, current_user) do
-    meeting
-    |> Auth.check!(current_user)
-    |> Repo.delete()
+    with {:ok, meeting} <- Auth.check(meeting, current_user), do: Repo.delete(meeting)
   end
 
   @doc """
   Builds a meeting to insert.
   """
   def new_meeting(current_user) do
-    new_meeting(current_user, [new_occurrence(current_user)])
+    with {:ok, occurrence} <- new_occurrence(current_user) do
+      new_meeting(current_user, [occurrence])
+    end
   end
 
   def new_meeting(current_user, occurrences) do
-    Auth.check!(
+    Auth.check(
       %Meeting{
         church_id: current_user.church.id,
         occurrences: occurrences
@@ -102,21 +104,20 @@ defmodule Blessd.Observance do
   Returns an `%Ecto.Changeset{}` for tracking meeting changes.
   """
   def change_meeting(%Meeting{} = meeting, current_user) do
-    meeting
-    |> Auth.check!(current_user)
-    |> Meeting.changeset(%{})
+    with {:ok, meeting} <- Auth.check(meeting, current_user) do
+      {:ok, Meeting.changeset(meeting, %{})}
+    end
   end
 
   @doc """
   Gets a single meeting occurrence.
-
-  Raises `Ecto.NoResultsError` if the MeetingOccurrence does not exist.
   """
-  def get_occurrence!(id, current_user) do
-    MeetingOccurrence
-    |> Auth.check!(current_user)
-    |> MeetingOccurrence.preload()
-    |> Repo.get!(id)
+  def find_occurrence(id, current_user) do
+    with {:ok, query} <- Auth.check(MeetingOccurrence, current_user) do
+      query
+      |> MeetingOccurrence.preload()
+      |> Repo.find(id)
+    end
   end
 
   @doc """
@@ -131,45 +132,46 @@ defmodule Blessd.Observance do
   end
 
   defp insert_occurrence(repo, %Meeting{id: meeting_id}, attrs, current_user) do
-    current_user
-    |> new_occurrence(meeting_id, nil)
-    |> MeetingOccurrence.changeset(attrs)
-    |> repo.insert()
+    with {:ok, occurrence} <- new_occurrence(current_user, meeting_id, nil) do
+      occurrence
+      |> MeetingOccurrence.changeset(attrs)
+      |> repo.insert()
+    end
   end
 
   @doc """
   Updates a meeting occurrence.
   """
   def update_occurrence(%MeetingOccurrence{} = occurrence, attrs, current_user) do
-    occurrence
-    |> Auth.check!(current_user)
-    |> MeetingOccurrence.changeset(attrs)
-    |> Repo.update()
+    with {:ok, occurrence} <- Auth.check(occurrence, current_user) do
+      occurrence
+      |> MeetingOccurrence.changeset(attrs)
+      |> Repo.update()
+    end
   end
 
   @doc """
   Deletes a Meeting occurrence.
   """
   def delete_occurrence(%MeetingOccurrence{} = occurrence, current_user) do
-    occurrence
-    |> Auth.check!(current_user)
-    |> Repo.delete()
+    with {:ok, occurrence} <- Auth.check(occurrence, current_user), do: Repo.delete(occurrence)
   end
 
   @doc """
   Returns an `%Ecto.Changeset{}` for tracking meeting occurrence changes.
   """
   def change_occurrence(%MeetingOccurrence{} = occurrence, current_user) do
-    occurrence
-    |> Auth.check!(current_user)
-    |> MeetingOccurrence.changeset(%{})
+    with {:ok, occurrence} <- Auth.check(occurrence, current_user) do
+      {:ok, MeetingOccurrence.changeset(occurrence, %{})}
+    end
   end
 
   @doc """
   Builds a meeting occurrence to insert.
   """
-  def new_occurrence(current_user, meeting_id \\ nil, date \\ Timex.today()) do
-    Auth.check!(
+  # TODO - we should consider user's timezone here
+  def new_occurrence(current_user, meeting_id \\ nil, date \\ Date.utc_today()) do
+    Auth.check(
       %MeetingOccurrence{
         church_id: current_user.church.id,
         meeting_id: meeting_id,
@@ -183,40 +185,41 @@ defmodule Blessd.Observance do
   Returns the list of people with their attendants preloaded.
   """
   def list_people(current_user) do
-    Person
-    |> Auth.check!(current_user)
-    |> Person.preload()
-    |> Person.order()
-    |> Repo.all()
+    with {:ok, query} <- Auth.check(Person, current_user) do
+      query
+      |> Person.preload()
+      |> Person.order()
+      |> Repo.list()
+    end
   end
 
   @doc """
   Search for people and preload their attendants.
   """
-  def search_people(query, current_user) do
-    Person
-    |> Auth.check!(current_user)
-    |> Person.preload()
-    |> Person.order()
-    |> Person.search(query)
-    |> Repo.all()
+  def search_people(query_str, current_user) do
+    with {:ok, query} <- Auth.check(Person, current_user) do
+      query
+      |> Person.preload()
+      |> Person.order()
+      |> Person.search(query_str)
+      |> Repo.list()
+    end
   end
 
   @doc """
   Creates an attendant if it does not exists and removes it if it exists.
   """
   def toggle_attendant(person_id, occurrence_id, current_user) do
-    case Repo.get_by(Attendant, person_id: person_id, meeting_occurrence_id: occurrence_id) do
-      nil ->
-        current_user
-        |> new_attendant()
-        |> Attendant.changeset(%{person_id: person_id, meeting_occurrence_id: occurrence_id})
-        |> Repo.insert()
+    case Repo.find_by(Attendant, person_id: person_id, meeting_occurrence_id: occurrence_id) do
+      {:ok, attendant} ->
+        with {:ok, attendant} <- Auth.check(attendant, current_user), do: Repo.delete(attendant)
 
-      %Attendant{} = attendant ->
-        attendant
-        |> Auth.check!(current_user)
-        |> Repo.delete()
+      {:error, :not_found} ->
+        with {:ok, attendant} <- new_attendant(current_user) do
+          attendant
+          |> Attendant.changeset(%{person_id: person_id, meeting_occurrence_id: occurrence_id})
+          |> Repo.insert()
+        end
     end
   end
 
@@ -224,6 +227,6 @@ defmodule Blessd.Observance do
   Builds an attendant to insert.
   """
   def new_attendant(current_user) do
-    Auth.check!(%Attendant{church_id: current_user.church.id}, current_user)
+    Auth.check(%Attendant{church_id: current_user.church.id}, current_user)
   end
 end
